Several things listed in description
set up necessary items for profile fragment, but am unable to fully implement due to a number of issues which will hopefully be resolved tomorrow.

rounded the profile picture images. can't actually retrieve or upload profile pictures w Parse right now because my changes to the database are not refreshing in Android Studio. need to ask a TA about this.